split Routes & Route up into Table, Route and Destination
note that Destination is the exact same thing as what a Route was.
It's the thing that sends data to one tcp endpoint, with spooling etc.
This commit is mostly just renames.
The Table is the global structure containing a list of Routes,
and Routes is basically a list of Endpoints.