define initial routing in cfg imperatively
after some more code cleanup, this should allow for less code,
because we can reuse the imperative command handling of the telnet admin
inputs.
the old way of declaratively defining the desired state requires more
structs and data juggling.
It's possible that I'm wrong about this and we'll revert back, but we'll
see.