implement redo spooling
i.e. the redo buffer will be spooled if we have spooling
and discarded otherwise. because if you restart your endpoint
and don't have spooling enabled, don't come to us to complain
that you lost some data, so it's easier to just drop in this case

note that we needed to increase some sleeps in the test :(
maybe related because the code is a bit more intense, or because
of the extra work logging?