new attempt/approach at verifying all received records with less allocations
as you can see, i couldn't make it work (esp at step 3 it gets tricky)
so disabled for now.  anyway with the previous approach all these checks
were fine, so it'll still be fine, and we check the num received anyway.
but we should try to reintroduce these checks somehow