better instrumentation + perf controls
* use go-metrics
expvar doesn't do aggregations we need, so use go-metrics.
also this lets us use a few more types.
we use a few custom patches for go-metrics, hence we currently point to
a fork, but there's PR's for all of them.
in particular, this version of go-metrics syncs its data to expvar,
which can be convenient.

* support sending metrics via carbon proto
* add more metrics, mostly wrt flushing perf

* expose important perf controls