fix Run() bugs + cleanup
* with default config, we would Run() a route twice,
first with 0 dests, than with N dests.
* from telnet interface -> Run() was never properly called with >0 dests,
  but the route would have the destination in its list and write to its
  nil channel, which would block the program

so redo how all that works, make sure we call Run() once for everything
and not more, because that would often spawn spurious goroutines that
would process data in parrallel.

Also no Run() method ever returns an error, so strip all that away.