Replace ogorek files that had non-ASCII names wtth ASCII ones
It appears that I somehow checked in a second set of ogorek.go and
ogorek_test.go with a different encoding for the second o that had an acute
accent. Meanwhile, the upstream package renamed the files to have ASCII names.

Remove both variants of the files with non-ASCII name, run
"go get github.com/kisielk/og-rek", and run "party" to add the files with
the ASCII names. Also ran tests.

Resolves #84