Add prepend and append options for destinations
This is an preliminary attempt to add support for prepending, and
appending, some arbitrary strings to the metrics name of before
relaying onward.

  I'm raising this to see if you are at all interested in this, or
if you think this should part of the relay at all. There will be
a minor performance impact when the settings are not used.

  Prepending should be reasonably efficient.

  Appending will create a fair bit of garbage in the current
implementation.

  I've not benchmarked either option yet, but can do if you are
interested in integrating something like this.

  I've also not added the web interface support yet.