rewrite before aggregator; fix race condition
it should always have been blacklist -> rewriter -> aggregator
that's also how it's documented.
in particular, this also fixes a race condition that may have caused bad
stats:
before this fix,
table sends a slice with the metric fields (key etc) to aggregator on a
channel, then invokes the rewriter which modifies the value in the first
slot of the array. when the aggregator reads field 0 (metric key) it is
concurrent with the rewriter modifying it, so the key seen might be pre
or post rewriting, and this is also confirmed via the race detector.

worse, this also caused panics when the key being changed caused the
regex expansion in the aggregator regex stuff to not find the needed
matching groups as shown in #197