make blocking/nonblocking controllable per route
(currently only implemented for kafkaMdm/grafanaNet routes)

A route can behave in two ways:
1) blocking: if it cannot ingest any more metrics (e.g.
   remote endpoint is down or too slow, and the route has no
   internal buffer, or the buffer is full) the route can block
   its Dispatch() call, effectively blocking the table,
   providing valuable feedback to clients sending data into
   the relay via backpressure, but in the process also
   preventing other routes from getting any inputs,
   since the table is blocked.
2) non-blocking: if it cannot ingest any more metrics, it can
   just discard metrics, returning Dispatch() early,
   so that the table keeps running (servicing any other routes
   that may be active and healthy)
   and clients sending data into the relay are none the wiser.

historically, all carbon routes were (and still are) non-blocking.
The assumption was you always had multiple routes and when one had
trouble it was never a reason to block any other routes.
There's a chance for data loss, but is mitigated via spooling and
the metrics buffer.
The kafkaMdm route also has a buffer and non-blocking dispatching.

Now we're more commonly seeing setups that have only 1 route,
so it makes more sense to provide backpressure
(and blocking the table in the process), to minimize data loss
- though only if you have smart clients that can deal with this.
The grafanaNet route actually is blocking, in contrast to
all other routes/destinations (including kafkaMdm).
Notably if you use >1 grafanaNet routes you can block the table when
buffer runs full. see #167

I think the best solution is to support both modes and leave the
configuration up to the user. They can choose the desired behavior
based on how many routes they have, and whether the clients can
deal with backpressure.