New aggregator: rate calculation aggregation (#207)
Create a new aggregator providing a rate of change for monotonic counters.
The aggregator work by computing the rate of change using the following formula:

```
rate = (Vnew - Vold)/(Tnew - Told)
```

Note:
The aggregator uses the actual timestamp (second precision) of the metric.
This means that if the aggregation window is larger than the number of samples, the rate should still be calculated properly.

Known Bugs:
The metrics values received are not re-ordered if received out of order. This has a limited impact if the time between metrics is slow (10s).

Issue: #207