make Handler more powerful
Give it access to the net.Conn when we have one.
This enables it to inspect it for logging and instrumentation,
opening the door for 3rd party projects that want to use our code
but have their own logging and metrics approach.

We still get to enjoy the code reuse optimization of having the same
io.Reader-based code path for handling both conns and packets,
but it is now an implementation detail rather than enforcing it via
the Handler function signature.

While we're at it, make logging around connections more consistent
and better.