separate releases vs inbetweens wrt package repo and docker tag
the main problem we solve here is that currently, people just install
the latest package, or run the latest docker image, and then run into
issues. We want to make it safer for them to run the relay.
It does pressure us a bit more to release more regularly to get fixes
out, which means we need to test unreleased code more, but that isn't
anything new.