report aggregator input more accurately.
We already had a standard counter which gets incremented whenever
(wallclock) a new metric comes in, and occassionally gets reported
on the rhythm of the stats reporter,
and it just reports the current value at whatever time it is.
This is useful to get a realtime overview to see how metrics
flow in compared to wallclock.

This adds a different metric, which shows buckets "logically":
it reports based on the timestamp of the data bucket:
after the wait has passed and wallclock has moved on,
whatever the count was for that bucket with that timestamp