separate configuration stuff from table
The table package should not reach out into configuration packages such
as cfg and imperatives
Instead the table is something that should be setup _from_ imperatives
(this was already the case) and _from_ the config
So, we move out the config bits from table and make it work with the
generic interface.
Just as we already tested table setup from imperatives, this will let us
test the same through cfg as well. (see next commit)

While we're at it:
* move imperative's table Interface and MockTable to the table package,
  so it can be used in various places
* cleaner "TableConfig" type
* move validation of badMetrics settings and its creation
  to table config and construction (that was the only case from
  the moved config stuff that accessed the table directly and couldn't
  be interface-ized)

One interesting thing is that the cfg imports imperatives. Why? simply
because your toml configuration may contain imperative commands that
need to be executed