treat absense of rules as an error
our configparser library will happily take
a file that contains just prose and no ini data at all,
and consider it a valid file, which would result in us
using the default aggregation rules, which may be a user
mistake we'd silently cover up.

So as a caller of it, be a bit more assertive.