Deploy new packages when tag is added (#506)
If you merge a new release into master and then add a new tag, the publish_release workflow fails. This is because packages aren't being deployed when a new tag is added. This PR adds the build/package/deploy steps to the publish_release workflow so correct packages are being deployed before the publish step.