Remove toki
Remove the toki module, replaced with a rewritten tokre.
This is a tokenizer based on regular expressions.
The API is similar, but not the same.
The tokre module is pulled into the same repo, as there are
not many users of the original code.
Fixes issue https://github.com/grafana/carbon-relay-ng/issues/551